requirements.txt
fastapi
uvicorn
supabase
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
sqlalchemy
psycopg2-binary

